HACKY: GraphicsWorker: use fences on GLES3
OK, so those GL.Finish()-calls cause poor performance, because
they require the full OpenGL queue to be flushed and completed.

So let's do a bit more targeted fix:

- We assume that OpenGL ES 2.0 hardware does not really require
  prover syncronization, as it's usually pretty simple hardware.
- However, some GPUS (like Tegra K1) have hardware-support for
  multiple GPU contexts, thus requiring proper syncronization.
- We assume that all GPUs requring proper synronization also
  support OpenGL ES 3.0; after all OpenGL ES 2.0 does not provide
  proper sychronization commands.
- The above assumption does not hold for desktop-hardware, so we
  need some different approach there.

If we encounter a setup where the above assumptions does not hold, we
should make a targeted fix for those; they will most likely expose a
sync-extension we can use.

These public OpenGL ES 2.0 extensions deal with sync/fences:
- OES_EGL_sync / KHR_fence_sync (seems to be parts of the same system,
  also included in EGL 1.5)
- APPLE_sync (might be useful if we encounter problems on old iPhones)
- NV_fence (supported on Tegra 2/3, but not on Tegra K1)

A final, last resort would be to be falling back to GL.Finish(). But
let's hope we don't have to do this, because this negatively affects
performance on devices that already have enough performance issues.