Fuse.Elements: do not pass dc.CullFace to blitter
The CullFace state should not be passed to the blitter here. The cache
will already be filled with rendering respecting the CullFace setting,
and if we set the CullFace-setting in the blitter to Front (or Both),
we end up culling the caching-rect itself, causing *all* rendering to
be dropped, even if it was rotated 180 degrees and should be
back-facing.

The important thing to realize here, is that the caching rectangle will
*always* be front-facing.

So let's just drop this. It's not going to do any good.